nwe Util updated phaseTwo
